requirements.txt
flask
requests
beautifulsoup4
rapidfuzz
python-dotenv
gunicorn